added session validation to server actions

+++ src/util/newspapers/actions.ts
@@ -1,16 +1,26 @@
 'use server'
 
+import { authOptions } from '@/app/api/auth/[...nextauth]/authOptions'
 import prisma from '@/lib/prisma'
 import { del } from '@vercel/blob'
+import { getServerSession } from 'next-auth'
 
 export const createNewspaper = async (newspaper: any) => {
-  'use server'
+  const session = await getServerSession(authOptions)
+  const user = session?.user
+  if (!user?.id || !user?.isAdmin) {
+    return
+  }
   await prisma.newspaper.create({
     data: newspaper
   })
 }
 export const editNewspaper = async (id: number, data: any) => {
-  'use server'
+  const session = await getServerSession(authOptions)
+  const user = session?.user
+  if (!user?.id || !user?.isAdmin) {
+    return
+  }
   const oldURL = data.oldURL
 
   if (oldURL) {
@@ -27,7 +37,11 @@ export const editNewspaper = async (id: number, data: any) => {
   })
 }
 export const deleteNewspaper = async (id: number) => {
-  'use server'
+  const session = await getServerSession(authOptions)
+  const user = session?.user
+  if (!user?.id || !user?.isAdmin) {
+    return
+  }
   await prisma.newspaper.delete({
     where: {
       id
